+++ src/components/Main/TitleSection.jsx
@@ -31,9 +31,9 @@ const IconFigcaption = styled.figcaption`
 function TitleSection() {
   return (
     <TitleContainer>
-      <TextGenerator text='Lets Lead The Industry To A Broader FIELD' font='Goblin One' />
+      <TextGenerator text="Let's Lead The Industry To A Broader FIELD" font='Goblin One' />
       <Figure $position='absolute'>
-        <Icon src='./../../public/scrollDown.png' />
+        <Icon src='scrollDown.png' />
         <IconFigcaption>아래로 스크롤하세요</IconFigcaption>
       </Figure>
     </TitleContainer>
