+++ src/pages/AboutPage.jsx
@@ -31,6 +31,7 @@ const H2 = styled.h2`
 `;
 
 const NanumH2 = styled(H2)`
+  font-size: 1.5rem;
   font-family: 'Nanum Myeongjo', serif;
   line-height: 1.3;
   word-break: keep-all;
@@ -146,9 +147,9 @@ function AboutPage() {
     2008: [{title: 'FIELD의 시작', get: '서울대학교, KAIST, POSTECH에서 학술 및 인적 교류 시작'}],
     2009: [{title: '첫 FIELD CAMP 개최', get: '고려대학교, 연세대학교 참여 시작'}],
     2016: [
-      {title: '대한산업공학회 산하공식단체 인준', get: '전국 대학교 학생 대상 캠프 주최 및 참여'},
+      {title: '대한산업공학회 산하 공식 단체 인준', get: '전국 대학교 학생 대상 캠프 주최 및 참여'},
     ],
-    2017: [{title: '전국단위 활동', get: '전국 단위 FIELD 활동'}],
+    2017: [{title: '전국 단위 활동', get: '전국 단위 FIELD 활동'}],
     2018: [{title: '고교방문 설명회 진행', get: '고등학생들을 대상으로 한 멘토링 진행'}],
     2022: [{title: 'FIELD 유튜브 개설', get: '산업공학 관련 영상 제작'}],
     2023: [{title: '코로나 이후 FIELD', get: '2년 만에 성황리에 개최된 FIELD CAMP'}],
@@ -159,7 +160,7 @@ function AboutPage() {
       <AccessibilityHidden>어바웃 필드</AccessibilityHidden>
       <TitleContainer>
         <NanumH2 margin='2rem 0'>전국 대학생 산업공학도 모임</NanumH2>
-        <GoblinP size='2.25rem' line='1.5'>
+        <GoblinP size='2.25rem' line='1.4'>
           <span>
             <FirstAlphabet color='red'>F</FirstAlphabet>uture
           </span>
@@ -176,10 +177,10 @@ function AboutPage() {
             <FirstAlphabet color='blue'>D</FirstAlphabet>reamers
           </span>
         </GoblinP>
-        <P line='1.5' margin='2rem 0 0 0' size='1rem'>
+        <P line='1.2' margin='2rem 0 0 0' size='1rem'>
           FIELD란, ‘Future Industrial Engineering Leaders and Dreamers’ 의 약자로, 미래의 핵심
-          리더들이 될 산업 공학도들이 모여 서로의 꿈과 비전, 생각 등을 공유할 수 있는 교류의 장을
-          만든다는 목표 아래 모인 &apos;전국 대학생 산업공학도 동아리&apos; 입니다.
+          리더들이 될 산업공학도들이 모여 서로의 꿈과 비전, 생각 등을 공유할 수 있는 교류의 장을
+          만든다는 목표 아래 모인 &apos;전국 대학생 산업공학도 모임&apos; 입니다.
         </P>
         <FlexCenter>
           <Figure position='absolute'>
